Add study screen

+++ app/src/main/java/jp/developer/bbee/englishmemory/domain/model/StudyStatus.kt
@@ -9,10 +9,10 @@ import androidx.room.Entity
 data class StudyStatus(
     val english: String,
     val wordType: String,
-    val numberOfQuestion: Int = 0,
-    val scoreRate: Int = 0,
-    val countMiss: Int = 0,
-    val countCorrect: Int = 0,
-    val isLatestAnswerCorrect: Boolean = false,
-    val isFavorite: Boolean = false,
+    val numberOfQuestion: Int = 0, // 出題回数
+    val scoreRate: Double = 0.0, // 正解率
+    val countMiss: Int = 0, // 間違えた回数
+    val countCorrect: Int = 0, // 正解した回数
+    val isLatestAnswerCorrect: Boolean = false, // 最新の回答が正解かどうか
+    val isFavorite: Boolean = false, // お気に入りかどうか
 )
